% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{call_api}
\alias{call_api}
\title{Call the API directly}
\usage{
call_api(endpoint, ...)
}
\arguments{
\item{endpoint}{Function name e.g. 'getConstituencies'}

\item{...}{\code{endpoint}'s (often optional) named arguments}
}
\value{
Response content, as unserialized by \code{jsonlite::fromJSON}
}
\description{
All the other functions call this one - it's exported only
for debugging purposes - so use them instead.
}
\details{
If you're really curious about implementation, read on.
Each API function introspects to see what its function name
is, bundles up its named arguments, and calls this function with them.
Consequently, aside from a bit of argument checking and/or return value
massaging, every function is implemented exactly the same way.
}

