% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{get_api_key}
\alias{get_api_key}
\title{Get the API key}
\usage{
get_api_key()
}
\value{
The current API key
}
\description{
Get the current API key or request it if not present.  When you provide a key
it will be stored as a local environment variable \code{TWFY_API_KEY} and also
in your \code{~/.Renviron}, making it available for all subsequent R sessions.
}
\details{
API keys can be requested at \url{http://theyworkforyou.com/api/key}.
}

