% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getBoundary}
\alias{getBoundary}
\title{Get constituency boundary as a KML file}
\usage{
getBoundary(name)
}
\arguments{
\item{name}{Name of constituency}
}
\value{
KML file
}
\description{
NOTE: This is not implemented.
}
