% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getComments}
\alias{getComments}
\title{Fetch comments left on TheyWorkForYou}
\usage{
getComments(start_date = NULL, end_date = NULL, search = NULL,
  pid = NULL, page = NULL, num = NULL)
}
\arguments{
\item{start_date}{Beginning date}

\item{end_date}{End date}

\item{search}{A search string}

\item{pid}{A person identifier}

\item{page}{which page of results to provide. Defaults to first page}

\item{num}{Number of results to return}
}
\value{
A list of user comments
}
\description{
Fetch comments left on TheyWorkForYou
}

