% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getCommittee}
\alias{getCommittee}
\title{Fetch members of a parliamentary committee}
\usage{
getCommittee(name = NULL, date = NULL)
}
\arguments{
\item{name}{Name of the committee. Partial names will be matched.}

\item{date}{Date for which the committee membership is required}
}
\value{
A list of committee members, and also committees if \code{name} is
        not specified.
}
\description{
If \code{name} is not specified, all committees and their members are
returned.
}
\details{
Theyworkforyou notes that "We have no information since the
2010 general election, and information before may be inaccurate."
}

