% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getConstituencies}
\alias{getConstituencies}
\title{Fetch UK parliament constituency names}
\usage{
getConstituencies(date = NULL, search = NULL)
}
\arguments{
\item{date}{Constituencies names as of this date}

\item{search}{Search string}
}
\value{
a list of elements each of which has a \code{name},
        or, if simplified, a character vector of constituency names.
}
\description{
Fetch UK parliament constituency names
}

