% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getConstituency}
\alias{getConstituency}
\title{Get information for a constituency}
\usage{
getConstituency(name = NULL, postcode = NULL)
}
\arguments{
\item{name}{Name of constituency}

\item{postcode}{A postcode

One of \code{name} or \code{postcode} is required.}
}
\value{
A data.frame with columns
       \itemize{
         \item{\code{name}}{Constituency name}
         \item{\code{pa_id}}{Constituency identifier}
         \item{\code{bbc_constituency_id}}{BBC numeric identifier}
         \item{\code{guardian_election_results}}{URL for election results}
         \item{\code{guardian_id}}{Guardian numeric identifier}
         \item{\code{guardian_name}}{Guardian name identifier}
       }
}
\description{
Get information for a constituency
}

