% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getDebates}
\alias{getDebates}
\title{Get information about parliamentary debates}
\usage{
getDebates(type = c("commons", "westminsterhall", "lords", "scotland",
  "northernireland"), date = NULL, search = NULL, person = NULL,
  gid = NULL, order = c("d", "r"), page = NULL, num = NULL)
}
\arguments{
\item{type}{location of the debate. Defaults to the House of Commons}

\item{date}{Date on which debates are required}

\item{search}{A search string}

\item{person}{A person identifier to specify which speaker's contributions are required}

\item{gid}{A speech/debate identifier to restrict to a particular debate}

\item{order}{whether to order results by date or relevance. Defaults to date}

\item{page}{which page of results to provide. Defaults to first page}

\item{num}{Number of results to return}
}
\value{
A complex data.frame of debate information.  Documentation is somewhat
        lacking, and many columns contain data.frames so you'll have to dig
        around.
}
\description{
Get information about parliamentary debates
}
