% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getGeometry}
\alias{getGeometry}
\title{Return geometry information for a constituency}
\usage{
getGeometry(name)
}
\arguments{
\item{name}{Name of constituency}
}
\value{
A list with constituency \code{name} and geometry information (see Details)
}
\description{
Returns a list with elements \itemize{
 \item{\code{parts}}{}
 \item{\code{area}}{}
 \item{\code{srid_n}}{}
 \item{\code{min_e}}{}
 \item{\code{centre_e}}{}
 \item{\code{max_e}}{}
 \item{\code{min_n}}{}
 \item{\code{centre_n}}{}
 \item{\code{max_n}}{}
 \item{\code{min_lat}}{}
 \item{\code{centre_lat}}{}
 \item{\code{max_lat}}{}
 \item{\code{min_long}}{}
 \item{\code{centre_long}}{}
 \item{\code{max_long}}{}
}
}

