% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getHansard}
\alias{getHansard}
\title{Search Hansard}
\usage{
getHansard(search = NULL, person = NULL, order = c("d", "r"),
  page = NULL, num = NULL)
}
\arguments{
\item{search}{A search string}

\item{person}{A person identifier}

\item{order}{whether to order results by date or relevance. Defaults to date}

\item{page}{which page of results to provide. Defaults to first page}

\item{num}{Number of results to return}
}
\value{
A list of search results
}
\description{
Search Hansard
}
