% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getLord}
\alias{getLord}
\title{Fetch information for a Lord}
\usage{
getLord(id)
}
\arguments{
\item{id}{Lord identifier}
}
\value{
List of a Lord's periods in the upper house.
}
\description{
Fetch information for a Lord
}

