% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getMLA}
\alias{getMLA}
\title{Fetch a member of the London Assembly}
\usage{
getMLA(id = NULL, postcode = NULL, constituency = NULL,
  always_return = NULL)
}
\arguments{
\item{id}{A MLA identifier}

\item{postcode}{A postcode, which specifies a region represented by the MLA}

\item{constituency}{Name of a constituency}

\item{always_return}{whether to try to return an MP even if
the seat is vacant or it before an election}
}
\value{
A list of information about an MLA
}
\description{
Fetch a member of the London Assembly
}

