% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getMLAs}
\alias{getMLAs}
\title{Fetch information about MLAs}
\usage{
getMLAs(date = NULL, party = NULL, search = NULL)
}
\arguments{
\item{date}{The date for which the MLAs are required}

\item{party}{Restrict MLAs to those in a party}

\item{search}{A search string}
}
\value{
A list of lists of MLA information
}
\description{
Fetch information about MLAs
}
