% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getMP}
\alias{getMP}
\title{Fetch information about an MP}
\usage{
getMP(id = NULL, postcode = NULL, constituency = NULL,
  always_return = NULL)
}
\arguments{
\item{id}{An MP identifier}

\item{postcode}{A postcode, used to identify a constituency and thereby an MP}

\item{constituency}{Name of a constituency}

\item{always_return}{whether to try to return an MP even if
the seat is vacant or it before an election}
}
\value{
A list of information about an MP
}
\description{
Fetch information about an MP
}

