% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getMPInfo}
\alias{getMPInfo}
\title{Fetch more information about a Member of Parliament}
\usage{
getMPInfo(id, fields = NULL)
}
\arguments{
\item{id}{An MP identifier}

\item{fields}{A comma separated character vector of field names}
}
\value{
a list
}
\description{
Unless you want all the four hundred or so fields available about an MP you should
probably specify the ones you want in a comma-separated string to \code{fields}.
}
\details{
Note that unlike other functions in this package the result is a list
}

