% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getMPs}
\alias{getMPs}
\title{Fetch Members of Parliament}
\usage{
getMPs(date = NULL, party = NULL, search = NULL)
}
\arguments{
\item{date}{Date for which the MP list is constructed}

\item{party}{Restrict to MPs in this party}

\item{search}{A search string}
}
\value{
A list of MPs or none if parliament is not in session.
}
\description{
Note that this will be the list as of today's date.  Before a general
election, there are technically no MPs so the list will be empty.
If this is not what you are expected, specify an earlier date.
}
