% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getMPsInfo}
\alias{getMPsInfo}
\title{Fetch more information about a Member of Parliament}
\usage{
getMPsInfo(ids, fields = NULL)
}
\arguments{
\item{ids}{A character vector of comma-separated MP identifiers}

\item{fields}{A comma separated character vector of field names}
}
\value{
A list of list of MP information
}
\description{
Fetch more information about a Member of Parliament
}

