% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getMSP}
\alias{getMSP}
\title{Fetch information on a Member of the Scottish Parliament}
\usage{
getMSP(id = NULL, postcode = NULL, constituency = NULL,
  always_return = NULL)
}
\arguments{
\item{id}{An MSP identifier}

\item{postcode}{A postcode, which specifies the constiuency whose MSP is required}

\item{constituency}{Name of a constituency}

\item{always_return}{whether to try to return an MP even if
the seat is vacant or it before an election}
}
\value{
A list of information about an MSP
}
\description{
Fetch information on a Member of the Scottish Parliament
}

