% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getMSPs}
\alias{getMSPs}
\title{Fetch information about Members of the Scottish Parliament}
\usage{
getMSPs(date = NULL, party = NULL, search = NULL)
}
\arguments{
\item{date}{The date for which the MSPs are required}

\item{party}{Restrict MSPs to those in a party}

\item{search}{A search string}
}
\value{
A list of lists of SP information
}
\description{
Fetch information about Members of the Scottish Parliament
}

