% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getPerson}
\alias{getPerson}
\title{Fetch information about a person}
\usage{
getPerson(id)
}
\arguments{
\item{id}{A person identifier}
}
\value{
A list of information about a person
}
\description{
Fetch information about a person
}

