% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getWMS}
\alias{getWMS}
\title{Fetch written ministerial responses}
\usage{
getWMS(date = NULL, search = NULL, person = NULL, gid = NULL,
  order = c("d", "r"), page = NULL, num = NULL)
}
\arguments{
\item{date}{Date for which responses are required}

\item{search}{A search string}

\item{person}{A person identifier to specify which minister provided the answers}

\item{gid}{A reponse identifier to return}

\item{order}{whether to order results by date or relevance. Defaults to date}

\item{page}{which page of results to provide. Defaults to first page}

\item{num}{Number of results to return}
}
\value{
A list of information about written ministerial responses
}
\description{
Fetch written ministerial responses
}
