% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the API key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{a new API key}
}
\value{
Nothing
}
\description{
This function manually sets or resets the API key.  It is mostly useful for
updating your key as all API function calls will ask for a key and store
it locally and in \code{~/.Renviron} any time they do not find one.
}

