% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy-package.R
\docType{package}
\name{twfy}
\alias{twfy}
\alias{twfy-package}
\title{twfy: A package for accessing TheyWorkForYou}
\description{
twfy is an R wrapper around the API of TheyWorkForYou, a parliamentary
monitoring site that scrapes and repackages Hansard (the UK's parliamentary
record) and augments it with information from the Register of Members'
Interests, election results, and voting records to provide a unified
source of information about UK legislators and their activities.                                                    <http://www.theyworkforyou.com> for details.
}
\details{
twfy functions currently mirror the structure and naming scheme of the API
as documented at \url{http://theyworkforyou.com/api/}. While the package
attempts where possible to provide more detail on structure of the API's
return values, this is not guaranteed to be correct and the link above
should be your primary source of documentation.
}
