% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{convertURL}
\alias{convertURL}
\title{Convert URL}
\usage{
convertURL(url)
}
\arguments{
\item{url}{url you want to convert}
}
\value{
A list containing \code{gid} and \code{url}.
}
\description{
Converts a parliament.uk Hansard URL into a
TheyWorkForYou one, if possible.
}
