% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getBoundary}
\alias{getBoundary}
\title{Get constituency boundary information as KML}
\usage{
getBoundary(name)
}
\arguments{
\item{name}{Name of constituency}
}
\value{
A character vector full of KML (an XML dialect)
}
\description{
Get constituency boundary information as KML
}
