% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getComments}
\alias{getComments}
\title{Get comments left on TheyWorkForYou}
\usage{
getComments(start_date = NULL, end_date = NULL, search = NULL,
  pid = NULL, page = NULL, num = NULL)
}
\arguments{
\item{start_date}{Beginning date}

\item{end_date}{End date}

\item{search}{A search string}

\item{pid}{A person identifier}

\item{page}{which page of results to provide. Defaults to first page}

\item{num}{Number of results to return}
}
\value{
A list containing data.frames with columns \itemize{
  \item{\code{comment_id}}{Comment identifier}
  \item{\code{user_id}}{User identifier}
  \item{\code{epobject_id}}{}
  \item{\code{body}}{Text of comment}
  \item{\code{posted}}{date and time posted}
  \item{\code{major}}{}
  \item{\code{gid}}{}
  \item{\code{firstname}}{Commenter's first name}
  \item{\code{lastname}}{Commenter's last name}
  \item{\code{url}}{URL of the comment}
  \item{\code{useurl}}{URL endpoint for commenter}
}
}
\description{
Get comments left on TheyWorkForYou
}

