% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getConstituencies}
\alias{getConstituencies}
\title{Fetch UK parliament constituency names}
\usage{
getConstituencies(date = NULL, search = NULL)
}
\arguments{
\item{date}{Date for which constituency names are required}

\item{search}{Search string}
}
\value{
a data.frame with single column \code{name} containing
        constituency names
}
\description{
Fetch UK parliament constituency names
}

