% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getConstituency}
\alias{getConstituency}
\title{Fetch a UK Parliament constituency}
\usage{
getConstituency(name = NULL, postcode = NULL)
}
\arguments{
\item{name}{Name of constituency}

\item{postcode}{A postcode

One of \code{name} or \code{postcode} is required.}
}
\value{
A list with elements including constituency \code{name}, and identifiers
}
\description{
Fetch a UK Parliament constituency
}

