% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getLords}
\alias{getLords}
\title{Fetch a list of Lords}
\usage{
getLords(date = NULL, party = NULL, search = NULL)
}
\arguments{
\item{date}{iso-style date, e.g. "1990-01-02", to compile a list of Lords for}

\item{party}{Include only Lords from this party}

\item{search}{A search term}
}
\value{
A data.frame of information about Lords
}
