% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getLords}
\alias{getLords}
\title{Fetch a list of Lords}
\usage{
getLords(date = NULL, party = NULL, search = NULL)
}
\arguments{
\item{date}{iso date, e.g. "1990-01-02", to compile the list for}

\item{party}{Include only Lords from this party}

\item{search}{A search term}
}
\value{
A List of Lords
}
\description{
Fetch a list of Lords
}
