% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getMLAs}
\alias{getMLAs}
\title{Get information on Members of the Legislative Assembly (Northern Ireland)}
\usage{
getMLAs(date = NULL, party = NULL, search = NULL)
}
\arguments{
\item{date}{The date for which the MLAs are required}

\item{party}{Restrict MLAs to those in a party}

\item{search}{A search string}
}
\value{
A data.frame with columns \itemize{
  \item{\code{member_id} }{Member identifier for each spell in the Legislative Assembly}
  \item{\code{person_id} }{Person identifier}
  \item{\code{name} }{Full name of MLA}
  \item{\code{party} }{Party represented}
  \item{\code{constituency} }{Name of MLA's constituency}
}
}
\description{
Get information on Members of the Legislative Assembly (Northern Ireland)
}

