% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getMPInfo}
\alias{getMPInfo}
\title{Fetch more information about a Member of Parliament}
\usage{
getMPInfo(id, fields = NULL)
}
\arguments{
\item{id}{An MP identifier}

\item{fields}{A comma separated character vector of field names}
}
\value{
A list of information about an MP
}
\description{
Unless you want all the hundred or so fields available about an MP you should
probably specify the ones you want in a comma-separated string to \code{fields}.
}

