% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getMSPs}
\alias{getMSPs}
\title{Get information about Members of the Scottish Parliament}
\usage{
getMSPs(date = NULL, party = NULL, search = NULL)
}
\arguments{
\item{date}{The date for which the MSPs are required}

\item{party}{Restrict MSPs to those in a party}

\item{search}{A search string}
}
\value{
A data.frame with columns \itemize{
  \item{\code{member_id} }{Member identifier for each spell in the Scottish parliament}
  \item{\code{person_id} }{Person identifier}
  \item{\code{name} }{Full name of MSP}
  \item{\code{party} }{Party represented}
  \item{\code{constituency} }{Name of MSP's constituency}
}
}
\description{
Get information about Members of the Scottish Parliament
}

