% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy.R
\name{getWrans}
\alias{getWrans}
\title{Get written answers to questions}
\usage{
getWrans(date = NULL, search = NULL, person = NULL, gid = NULL,
  order = c("d", "r"), page = NULL, num = NULL)
}
\arguments{
\item{date}{Date for which answers are required}

\item{search}{A search string}

\item{person}{A person identifier to specify who provided the answers}

\item{gid}{A written question and answer identifier to return}

\item{order}{whether to order results by date or relevance. Defaults to date}

\item{page}{which page of results to provide. Defaults to first page}

\item{num}{Number of results to return}
}
\value{
A list of written answers information
}
\description{
Get written answers to questions
}
