% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfy-package.R
\docType{package}
\name{twfy}
\alias{twfy}
\alias{twfy-package}
\title{twfy: A package for accessing TheyWorkForYou}
\description{
This package is a light wrapper around TheyWorkForYou's API,
allowing queries using the verbs described at
\url{http://theyworkforyou.com/api/}.  This should be your primary
source of documentation as the package develops.
}

